+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Move homebrew formula from github. (@miry)
+
 ## [0.8.1] - 2022-08-18
 ### Changed
 - Medium: use field `webCanonicalUrl` for `canonical_url`. (@miry)